requirements.txt
selenium
flask
google-generativeai
python-dotenv
